backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=21.3.0",
    "black>=23.0.0",
    "cryptography>=41.0.0",
    "django>=4.2.0,<5.0.0",
    "django-cors-headers>=4.0.0",
    "django-filter>=25.1",
    "djangorestframework>=3.14.0",
    "flake8>=6.0.0",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
    "pyjwt>=2.8.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-django>=4.5.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "reportlab>=4.0.0",
    "sqlparse>=0.4.4",
]
